Updated with QR parser & Rekognition detector

+++ src/main/java/zxingdemo/Point.java
@@ -1,22 +1,22 @@
 package zxingdemo;
 
 public class Point {
-    int x;
-    int y;
+    float x;
+    float y;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
